Munch: Enable debug.sf.defer_refresh_rate_when_off
Since we have a video mode panel, we need to have this prop enabled
because the scheduler sets refresh rate to 120 Hz _after_ turning
screen off at 60 Hz (VRR idle) but the HWC ignores this and when
the display is turned back on SF thinks we are back at 120 Hz
while HWC is still rendering at 60 Hz. Defer setting refresh rate
until the screen is turned back on, to avoid this confusion.

Signed-off-by: peternmuller <pm563838@gmail.com>